requirements.txt
# Core deep learning packages
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning utilities
scikit-learn>=1.3.0

# Progress bars and utilities
tqdm>=4.65.0
PyYAML>=6.0.0

# System monitoring
psutil>=5.9.0

# Development and testing (optional)
jupyter>=1.0.0
pytest>=7.0.0

# Configuration management
python-dotenv>=1.0.0

# Additional utilities for model development
tensorboard>=2.13.0
wandb>=0.15.0